add subscribed stream completion
this needs an oauth token... need to write a script for that.

also, json parsing for this cannot easily be done with string
manipulation, which is why we use a small call to python for it. it's
not the best solution, but since livestreamer is itself a python
application, and we only use standard modules, it should be fine.